+++ lib/generators/dry/struct/struct_generator.rb
@@ -2,7 +2,7 @@ require 'generators/dry'
 
 module Dry
   module Generators
-    class ValidationGenerator < Base
+    class StructGenerator < Base
       source_root File.expand_path('templates', __dir__)
 
       def check_requirements
@@ -14,24 +14,24 @@ module Dry
         @class_name = "#{@name.camelize}"
       end
 
-      def copy_validator
-        path = File.join(VALIDATIONS_PATH, "#{@name}.rb")
-        template 'validation_template.rb', path
+      def copy_struct
+        path = File.join(STRUCTS_PATH, "#{@name}.rb")
+        template 'struct_template.rb', path
       end
 
       def copy_spec
         if Dir.exist? File.join(Rails.root, 'spec')
-          spec_path =  File.join(SPEC_PATH, 'validations')
+          spec_path =  File.join(SPEC_PATH, 'structs')
           path = File.join(spec_path, "#{@name}_spec.rb")
-          template 'validation_spec_template.rb', path
+          template 'struct_spec_template.rb', path
         end
       end
 
       def copy_test
         if Dir.exist? File.join(Rails.root, 'test')
-          test_path =  File.join(TEST_PATH, 'validations')
+          test_path =  File.join(TEST_PATH, 'structs')
           path = File.join(test_path, "#{@name}_test.rb")
-          template 'validation_test_template.rb', path
+          template 'struct_test_template.rb', path
         end
       end
     end