+++ lib/generators/dry/struct/templates/struct_template.rb
@@ -1,0 +1,5 @@
+module Structs
+  class <%= @class_name %> < Dry::Struct
+    # attribute :age, Types::Integer
+  end
+end