+++ CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2020-02-15
+## [0.3.0] - 2020-02-16
+### Added
+- A generator for dry-struct.
+- test/dummy/test/.gitkeep
+
+### Changed
+- Bumped version.
+- Moved path variables to module constants.
+- Changed namespace for generators.
+- Changed 'validators' to 'validations' to more align with dry-rb.
+
+### Removed
+- test/dummy/lib/generators directory.
+- test/dummy/test/lib directory.
+
+## [0.2.0] - 2020-02-16
 ### Added
 - This very nice changelog you are reading.
 - A spec directory for the dummy application.